move autoUpgrade to commands directory
