--- packages/gluestick/src/commands/autoUpgrade.js
@@ -1,8 +1,0 @@
-/* @flow */
-import type { Context } from '../types';
-
-const autoUpgrade = require('../autoUpgrade/autoUpgrade');
-
-module.exports = async (context: Context) => {
-  await autoUpgrade(context);
-};
