+++ packages/gluestick/src/commands/autoUpgrade/__tests__/autoUpgrade.test.js
