+++ packages/gluestick/src/commands/autoUpgrade/__tests__/updateDependencies.test.js
