+++ packages/gluestick/src/commands/autoUpgrade/__tests__/utils.test.js
