+++ packages/gluestick/src/commands/autoUpgrade/getSingleEntryFromGenerator.js
