+++ packages/gluestick/src/commands/autoUpgrade/index.js
@@ -1,0 +1,8 @@
+/* @flow */
+import type { Context } from '../../types';
+
+const autoUpgrade = require('./autoUpgrade');
+
+module.exports = async (context: Context) => {
+  await autoUpgrade(context);
+};
